+++ packages/tiers/src/index.ts
@@ -10,6 +10,7 @@ export {
   detectChallengeType,
   hasAkamaiChallenge,
   hasDdosGuardChallenge,
+  hasAwsWafChallenge,
   hasHcaptcha,
   hasImpervaChallenge,
   hasRecaptcha,